feat: добавить joi бля обработки запросов

+++ controllers/deleteCard.js
@@ -1,19 +1,24 @@
-import { NotFoundError } from '../errors/errors.js';
-import Card from '../models/card.js';
+import { NotFoundError, accessError } from "../errors/errors.js";
+import Card from "../models/card.js";
 
 export const deleteCard = (req, res, next) => {
   try {
-    Card.findByIdAndDelete(req.params.cardId)
+    Card.findById(req.params.cardId)
       .then((card) => {
-        if (!card) {
-           throw new NotFoundError("карточка с данным id не найдена");
+        if (card.owner === req.user) {
+          return Card.findByIdAndDelete(req.params.cardId)
+            .then((card) => {
+              if (!card) {
+                throw new NotFoundError("карточка с данным id не найдена");
+              }
+              res.status(200).json(card);
+            })
+            .catch(next);
         }
-        res.status(200).json(card);
+        throw new accessError("доступ отсутствует");
       })
       .catch(next);
-  } catch (error) {
-    return res.status(500).send({
-      message: 'На сервере произошла ошибка',
-    });
+  } catch {
+    next();
   }
 };
